Fix/Workaround: Accept ISO8601 Time format including Z timezone
ISO8601 datetime formatting requires a timezone. As a first workaround, I added a 'Z' for the "Zulu" (GMT) timezone.

In ISO8601 other timezones are also alloweed, eg,. by using a "+0100"  instead of the Z. This means that the validator should also allow it. 

Also, the fraction of seconds  is optional (`.%f `)